pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sneaky-helper"
version = "0.2.0"
dependencies = [
    "aiohttp~=3.8.6",
    'aiosignal~=1.3.1',
    "async-timeout~=4.0.3",
    "attrs~=23.1.0",
    "cffi~=1.16.0",
    "charset-normalizer~=3.3.2",
    "discord~=2.3.2",
    "discord.py~=2.3.2",
    "ffmpeg-python~=0.2.0",
    "frozenlist~=1.4.0",
    "future~=0.18.3",
    "idna~=3.4",
    "imageio-ffmpeg~=0.4.9",
    "multidict~=6.0.4",
    "numpy~=1.26.3",
    "py-cord~=2.4.1",
    "pycord~=0.1.1",
    "pycparser~=2.21",
    "PyNaCl~=1.5.0",
    "python-dotenv~=1.0.0",
    "yarl~=1.9.2",
]
requires-python = ">=3.8"
description = "Discord.py bot leveraging Whisper AI for transcription of voice chats"

[project.scripts]
sneaky-helper = "sneaky_helper.main:main"

[project.urls]
changelog = "https://github.com/hayesraven/sneaky-helper/blob/main/CHANGELOG.md"
issues = "https://github.com/hayesraven/sneaky-helper/issues"
repository = "https://github.com/hayesraven/sneaky-helper"

[project.optional-dependencies]
build = ["build >= 1.2"]
test = ["pytest ~= 8.1"]
